Finalizing repositories, services and configs
