+++ src/main/java/com/example/gnap/as/config/JacksonConfig.java
@@ -1,0 +1,28 @@
+package com.example.gnap.as.config;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * Jackson configuration for the GNAP Authorization Server.
+ * This class provides the ObjectMapper bean used for JSON serialization/deserialization.
+ */
+@Configuration
+public class JacksonConfig {
+
+    /**
+     * Configure ObjectMapper for JSON serialization/deserialization.
+     *
+     * @return the configured ObjectMapper
+     */
+    @Bean
+    public ObjectMapper objectMapper() {
+        ObjectMapper objectMapper = new ObjectMapper();
+        objectMapper.registerModule(new JavaTimeModule());
+        objectMapper.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+        return objectMapper;
+    }
+}
